requirements.txt
ipykernel==7.1.0
pandas==2.3.3
matplotlib==3.8.0
plotly==6.5.0
nbformat==5.10.4
lightweight-charts==2.1
yfinance==1.0
ipython==8.18.0
pyarrow==21.0.0
pyspark==3.5.0
minio==7.2.20
python-dotenv=1.2.1